requirements.txt
flask
ultralytics
opencv-python-headless
pillow
numpy
gunicorn
flask-cors